requirements.txt
anthropic-bedrock
boto3
openai
python-dotenv
